feat(interaction): implement InteractionQueryService
