+++ src/main/java/org/nfactorial/newsfeed/domain/interaction/service/FollowService.java
@@ -1,13 +1,9 @@
 package org.nfactorial.newsfeed.domain.interaction.service;
 
-import java.util.List;
-
 import org.nfactorial.newsfeed.common.code.ErrorCode;
 import org.nfactorial.newsfeed.common.exception.BusinessException;
-import org.nfactorial.newsfeed.domain.interaction.dto.response.FollowStatusResponse;
 import org.nfactorial.newsfeed.domain.interaction.entity.Follow;
 import org.nfactorial.newsfeed.domain.interaction.repository.FollowRepository;
-import org.nfactorial.newsfeed.domain.profile.dto.ProfileSummaryDto;
 import org.nfactorial.newsfeed.domain.profile.entity.Profile;
 import org.nfactorial.newsfeed.domain.profile.service.ProfileServiceApi;
 import org.springframework.stereotype.Service;
@@ -53,24 +49,4 @@ public class FollowService {
 		follower.decrementFollowCount();
 	}
 
-	@Transactional(readOnly = true)
-	public FollowStatusResponse checkFollowStatus(Long followerId, Long followingId) {
-
-		if (followerId.equals(followingId)) {
-			throw new BusinessException(ErrorCode.CANNOT_FOLLOW_SELF);
-		}
-
-		// profileService의 메소드에 id 존재여부 검증 위임, 별도의 서비스 계층 exists 메소드 생성 x 목적
-		Profile follower = profileService.getProfileById(followerId);
-		Profile following = profileService.getProfileById(followingId);
-
-		return FollowStatusResponse.of(followRepository.existsByFollowerAndFollowing(follower, following));
-	}
-
-	@Transactional(readOnly = true)
-	public List<ProfileSummaryDto> getFollowingProfiles(Long followerId) {
-
-		List<Long> followingIds = followRepository.findFollowingIdsByFollowerId(followerId);
-		return profileService.findProfileSummariesByIds(followingIds);
-	}
 }
