+++ src/main/java/org/nfactorial/newsfeed/domain/interaction/service/InteractionQueryService.java
@@ -1,0 +1,55 @@
+package org.nfactorial.newsfeed.domain.interaction.service;
+
+import java.util.List;
+
+import org.nfactorial.newsfeed.common.code.ErrorCode;
+import org.nfactorial.newsfeed.common.exception.BusinessException;
+import org.nfactorial.newsfeed.domain.interaction.dto.response.FollowStatusResponse;
+import org.nfactorial.newsfeed.domain.interaction.repository.FollowRepository;
+import org.nfactorial.newsfeed.domain.interaction.repository.LikeRepository;
+import org.nfactorial.newsfeed.domain.post.service.PostService;
+import org.nfactorial.newsfeed.domain.profile.dto.ProfileSummaryDto;
+import org.nfactorial.newsfeed.domain.profile.entity.Profile;
+import org.nfactorial.newsfeed.domain.profile.service.ProfileServiceApi;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import lombok.RequiredArgsConstructor;
+
+@Service
+@RequiredArgsConstructor
+@Transactional(readOnly = true)
+public class InteractionQueryService implements InteractionQueryServiceApi {
+
+	private final FollowRepository followRepository;
+	private final LikeRepository likeRepository;
+	private final ProfileServiceApi profileService;
+	private final PostService postService;
+
+	@Override
+	public FollowStatusResponse checkFollowStatus(Long followerId, Long followingId) {
+
+		if (followerId.equals(followingId)) {
+			throw new BusinessException(ErrorCode.CANNOT_FOLLOW_SELF);
+		}
+
+		// profileService의 메소드에 id 존재여부 검증 위임, 별도의 서비스 계층 exists 메소드 생성 x 목적
+		Profile follower = profileService.getProfileById(followerId);
+		Profile following = profileService.getProfileById(followingId);
+
+		return FollowStatusResponse.of(followRepository.existsByFollowerAndFollowing(follower, following));
+	}
+
+	@Override
+	public List<ProfileSummaryDto> getFollowingProfiles(Long followerId) {
+
+		List<Long> followingIds = followRepository.findFollowingIdsByFollowerId(followerId);
+		return profileService.findProfileSummariesByIds(followingIds);
+	}
+
+	@Override
+	public boolean hasLikedPost(Long postId, Long profileId) {
+
+		return likeRepository.existsByPostIdAndProfileId(postId, profileId);
+	}
+}
